+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_global_state/visitor.dart
@@ -9,10 +9,14 @@ class _Visitor extends RecursiveAstVisitor<void> {
   void visitVariableDeclaration(VariableDeclaration node) {
     super.visitVariableDeclaration(node);
 
-    if (!node.isFinal &&
-        !node.isConst &&
-        node.declaredElement?.enclosingElement is CompilationUnitElement) {
-      _declarations.add(node);
+    if (node.declaredElement?.enclosingElement is CompilationUnitElement) {
+      if (!node.isFinal && !node.isConst) {
+        _declarations.add(node);
+      }
+    } else {
+      if (node.declaredElement?.isStatic ?? false) {
+        _declarations.add(node);
+      }
     }
   }
 }