+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_global_state/avoid_global_state_rule_test.dart
@@ -25,13 +25,17 @@ void main() {
 
       RuleTestHelper.verifyIssues(
         issues: issues,
-        startLines: [1, 2],
-        startColumns: [5, 5],
+        startLines: [1, 2, 9, 10],
+        startColumns: [5, 5, 15, 18],
         locationTexts: [
           'answer = 42',
           'evenNumbers = [1, 2, 3].where((element) => element.isEven)',
+          'a',
+          'c',
         ],
         messages: [
+          'Avoid using global variable without const or final keywords.',
+          'Avoid using global variable without const or final keywords.',
           'Avoid using global variable without const or final keywords.',
           'Avoid using global variable without const or final keywords.',
         ],